added a nullifier event listener
