+++ src/boilerplate/common/commitment-storage.mjs
@@ -9,34 +9,12 @@ import logger from './logger.mjs';
 import utils from "zkp-utils";
 import { poseidonHash } from './number-theory.mjs';
 import { generateProof } from './zokrates.mjs';
-import { SumType, reduceTree, toBinArray, poseidonConcatHash } from './smt_utils.mjs';
-import { hlt } from './hash-lookup.mjs';
+import { temporaryUpdateNullifier, getnullifierMembershipWitness } from './nullifier-storage.mjs'
+
 
 const { MONGO_URL, COMMITMENTS_DB, COMMITMENTS_COLLECTION } = config;
 const { generalise } = gen;
 
-const TRUNC_LENGTH = 32; // Just for testing so we don't make more than 32 deep smt trees.
-const WHOLE_STATES = [WHOLE_STATE_NAMES];
-// structure for SMT
-const Branch = (leftTree, rightTree) => ({
-  tag: 'branch',
-  left: leftTree,
-  right: rightTree,
-});
-const Leaf = val => ({
-  tag: 'leaf',
-  val: val,
-});
-
-const SMT = SumType([Branch, Leaf], () => {
-  throw new TypeError('Invalid data structure provided');
-});
-
-let smt_tree = SMT(hlt[0]);
-let temp_smt_tree = SMT(hlt[0]); // for temporary updates before proof generation
-
-// Gets the hash of a smt_tree (or subtree)
-export const getHash = tree => reduceTree(poseidonConcatHash, tree);
 
 // function to format a commitment for a mongo db and store it
 export async function storeCommitment(commitment) {
@@ -128,43 +106,6 @@ export async function updateCommitment(commitment, updates) {
   return db.collection(COMMITMENTS_COLLECTION).updateOne(query, update);
 }
 
-// This is a helper function to insertLeaf in smt that calls the recursion
-function _insertLeaf(val, tree, binArr){
-	if (binArr.length > 0) {
-	  switch (tree.tag) {
-		case 'branch': // Recursively enter developed subtree
-		  return binArr[0] === '0'
-			? Branch(_insertLeaf(val, tree.left, binArr.slice(1)), tree.right)
-			: Branch(tree.left, _insertLeaf(val, tree.right, binArr.slice(1)));
-  
-			case 'leaf': // Open undeveloped subtree
-			return binArr[0] === '0'
-			  ? Branch(
-				  _insertLeaf(val, Leaf(hlt[TRUNC_LENGTH - (binArr.length - 1)]), binArr.slice(1)),
-				  Leaf(hlt[TRUNC_LENGTH - (binArr.length - 1)]),
-				)
-			  : Branch(
-				  Leaf(hlt[TRUNC_LENGTH - (binArr.length - 1)]),
-				  _insertLeaf(val, Leaf(hlt[TRUNC_LENGTH - (binArr.length - 1)]), binArr.slice(1)),
-				);
-	
-  
-		default: {
-		  return tree;
-		}
-	  }
-	} else return Leaf(val);
-  };
-  
-  // This inserts a value into the smt as a leaf
-  function insertLeaf(val, tree) {
-	const binArr = toBinArray(generalise(val))
-	const padBinArr = Array(254 - binArr.length)
-		.fill("0")
-		.concat(...binArr).slice(0, TRUNC_LENGTH);
-	return _insertLeaf(val, tree, padBinArr);
-  };
-
 // function to mark a commitment as nullified for a mongo db and update the nullifier tree
 export async function markNullified(commitmentHash, secretKey = null) {
     const connection = await mongo.connection(MONGO_URL);
@@ -184,9 +125,7 @@ export async function markNullified(commitmentHash, secretKey = null) {
 		nullifier: generalise(nullifier).hex(32),
       },
     };
-	// updating the original tree
-	await updateNullifierTree();
-
+	
     return db.collection(COMMITMENTS_COLLECTION).updateOne(query, update);
   }
 
@@ -463,94 +402,6 @@ export async function joinCommitments(
 	});
 
 	return { tx };
-}
-
-
- // This is a helper function for checkMembership
-const _getnullifierMembershipWitness = (binArr, element, tree, acc) => {
-	switch (tree.tag) {
-	  case 'branch':
-		return binArr[0] === '0'
-		  ? _getnullifierMembershipWitness(
-			  binArr.slice(1),
-			  element,
-			  tree.left,
-			  [getHash(tree.right) ].concat(acc),
-			)
-		  : _getnullifierMembershipWitness(
-			  binArr.slice(1),
-			  element,
-			  tree.right,
-			  [getHash(tree.left)].concat(acc),
-			);
-			case "leaf": {
-				if (binArr.length > 0) {
-					while (binArr.length > 0) {
-						binArr[0] === "0"
-							? (acc = [hlt[TRUNC_LENGTH - (binArr.length - 1)]].concat(acc))
-							: (acc = [hlt[TRUNC_LENGTH - (binArr.length - 1)]].concat(acc));
-						binArr = binArr.slice(1);
-					}
-					return { isMember: false, path: acc };
-				} else {
-					return tree.val !== element
-						? { isMember: false, path: acc }
-						: { isMember: true, path: acc };
-				}
-			}
-	  default:
-		return tree;
-	}
-  };
-
-export async function updateNullifierTree() {
-	smt_tree = 	temp_smt_tree;
-}
-
-export async function resetTemporaryNullifierTree() {
-	temp_smt_tree = smt_tree;
-}
-
-
-export function getnullifierMembershipWitness(nullifier) {
-
-	const binArr = toBinArray(generalise(nullifier))
-	const padBinArr = Array(254 - binArr.length)
-		.fill("0")
-		.concat(...binArr).slice(0, TRUNC_LENGTH);
-	const  membershipPath = _getnullifierMembershipWitness(padBinArr, nullifier, temp_smt_tree, []);
-    const root = getHash(temp_smt_tree);
-	const witness = {path : membershipPath.path, root: root}
-	return witness;
-
-}
-
-export async function temporaryUpdateNullifier(nullifier){
-	
-	temp_smt_tree = insertLeaf(generalise(nullifier).hex(32), temp_smt_tree);
-	
-}
-
-export async function reinstateNullifiers() {
-	const initialised = [];
-	const nullifiedCommitments = await getNullifiedCommitments();
-	if (!nullifiedCommitments) {
-		logger.info('No nullifiers to add to the tree');
-		return;
-	}
-	logger.warn(
-	'Reinstatiating nullifiers - NOTE that any nullifiers added from another client may not be known here, so the tree will be out of sync.',
-	);
-	for (const c of nullifiedCommitments) {
-		if (WHOLE_STATES.includes(c.name) && !initialised.includes(c.preimage.stateVarId)) {
-			logger.debug(`initialising state ${c.name}`);
-			smt_tree = insertLeaf(poseidonHash([BigInt(c.preimage.stateVarId), BigInt(0), BigInt(0)]).hex(32), smt_tree);
-			initialised.push(c.preimage.stateVarId);
-		}
-		logger.debug(`nullifying state ${c.name}: ${c.nullifier}`);
-		smt_tree = insertLeaf(c.nullifier, smt_tree);
-	}
-	temp_smt_tree = smt_tree;
 }
